Refactor authentication services: consolidate login and registration logic into authService and update imports accordingly
