+++ src/pages/Login/components/Login.tsx
@@ -4,7 +4,7 @@ import { SectionTitle } from '@/components/SectionTitle'
 import { SubmitButton } from '@/components/SubmitButton'
 import { TextInput } from '@/components/TextInput'
 import { useSession } from '@/hooks/useSession'
-import { loginUser } from '@/services/loginUser'
+import { loginUser } from '@/services/authService'
 import { FormDataType } from '@/types/dataTypes'
 import { formatFormData } from '@/utilities/formatFormData'
 import { useNavigate } from 'react-router-dom'
