+++ src/pages/Register/components/Register.tsx
@@ -2,7 +2,7 @@ import { LinkedAnotation } from '@/components/LinkedAnotation'
 import { SectionTitle } from '@/components/SectionTitle'
 import { SubmitButton } from '@/components/SubmitButton'
 import { TextInput } from '@/components/TextInput'
-import { registerUser } from '@/services/registerUser'
+import { registerUser } from '@/services/authService'
 import { FormDataType } from '@/types/dataTypes'
 import { formatFormData } from '@/utilities/formatFormData'
 import React, { useState } from 'react'
